tools/repomind/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "repomind"
version = "0.1.0"
description = "Deterministic repo-to-Obsidian-vault sync for claude-hangar workflows"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Giorgo Lazaridis", email = "kontakt@lazaridis.me" }]
keywords = ["obsidian", "vault", "sync", "claude-hangar"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: POSIX",
]
dependencies = [
    "typer>=0.12.0",
    "rich>=13.0.0",
    "pyyaml>=6.0.0",
    "pathspec>=0.12.0",
]

[project.scripts]
repomind = "repomind.cli:app"

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.5.0",
    "mypy>=1.10.0",
    "types-PyYAML>=6.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/repomind"]

[tool.hatch.build.targets.sdist]
include = [
    "src/repomind",
    "templates",
    "README.md",
    "pyproject.toml",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = [
    "--strict-markers",
    "--tb=short",
    "--cov=src/repomind",
    "--cov-report=term-missing",
    "--cov-fail-under=80",
]

[tool.coverage.run]
source = ["src/repomind"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "SIM", # flake8-simplify
    "N",   # pep8-naming
    "RUF", # ruff-specific
]
ignore = [
    "E501",  # line too long — handled by formatter
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["B011", "N802"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true

[[tool.mypy.overrides]]
module = ["pathspec.*"]
ignore_missing_imports = true
